feature brand

+++ src/backend/Domain/Shared/PagingResult.cs
@@ -11,9 +11,9 @@ namespace Domain.Shared
         public int PageIndex { get; private set; }
         public int PageSize { get; private set; }
         public int PageNumber { get { return PageIndex + 1; } set { PageIndex = value - 1; } }
-        public int TotalItems { get; private set; }
-        public int TotalPages { get { var total = TotalItems / PageSize; if (TotalItems % PageSize > 0) { total++; } return total; } private set { } }
-        public PagingResult(T data, int pageNumber, int pageSize, int totalItems) : base(data)
+        public long TotalItems { get; private set; }
+        public int TotalPages { get { var total = TotalItems / PageSize; if (TotalItems % PageSize > 0) { total++; } return (int)total; } private set { } }
+        public PagingResult(T data, int pageNumber, int pageSize, long totalItems) : base(data)
         {
             PageNumber = pageNumber;
             PageSize = pageSize;
